//Commit 8 Minor changes to the design/layout. Was not able to implement location services. Freezes the page for some reason. UpdatesActivity.java also doesn't do anything since there are no location services/Twitter API.
Potential Changes later on:
-Access List of countries better.
-Add Twitter API to show tweets from the heads of the countries
-Add maps/graphs of COVID-19
-Make the design a little better

Not bad for 3 days of work

//Commit 7
Did research on Twitter and Google Maps API. Got the very basics of Google Maps API down. Was able to implement it in CountriesActivity.java and HomeActivity.java. Twitter API needs more time to learn. Had to hardcode World COVID-19 data. It is in an object not an array. Need to use Volley to extract the data, but not enough time.

//Commit 6
Learned how to implement APIs. Added Death, Infection, Recovery rates for 49 countries. Not the most accurate, but works somewhat.

//Commit 5
Fixed UpdatesActivity.java from crashing. Added an expandable layout to StaySafeActivity. Added all countries to CountriesActivity. Still need to add APIs for Google Maps and Twitter feeds.

//Commit 4
UpdatesActivity.java is still crashing. Added Design to each page with ideas of what features need to be implemented. Planning on expandable layout for activity_staysafe.xml

//Commit 3
Added navigation to the other pages. UpdatesActivity.java page is crashing for some unknown reason.

//Commit 2
Added a bottom navigation to the activity_home.xml that switches to different pages

//Commit 1
Renamed files. Added labels to HomeActivity.java